moved slider to shared
